Text_2_SQL/requirements.txt
streamlit
google-generativeai
python-dotenv


